1.1.0
- Added pokemon finder functionality similar to dexnav.
At the moment it can show the best Pokémon in the neighborhood and show the coordinates of the place where it was found. In the next patch I plan to add an option to pokenav to track the found pokémon.
Please note that currently no Pokemon can have an Egg Move or Hidden Ability without additional side mods. I advise you to install Cobblemon Counter and Cobblemon Unchained if you have not already done so
- Impoved spawn checking feature:  now regional forms are displayed correctly, added check for the "implemented" parameter, added check for presence of "not_modeled" label (temporary solution for compatibility with Gravelmon)
- Minor changes to the display of some UI elements